# Copy-on-write for record values