--- a/crates/nu-cli/src/completions/variable_completions.rs
+++ b/crates/nu-cli/src/completions/variable_completions.rs
@@ -68,9 +68,7 @@ impl Completer for VariableCompletion {
                         self.var_context.1.clone().into_iter().skip(1).collect();
 
                     if let Some(val) = env_vars.get(&target_var_str) {
-                        for suggestion in
-                            nested_suggestions(val.clone(), nested_levels, current_span)
-                        {
+                        for suggestion in nested_suggestions(val, &nested_levels, current_span) {
                             if options.match_algorithm.matches_u8_insensitive(
                                 options.case_sensitive,
                                 suggestion.suggestion.value.as_bytes(),
@@ -117,8 +115,7 @@ impl Completer for VariableCompletion {
                     nu_protocol::NU_VARIABLE_ID,
                     nu_protocol::Span::new(current_span.start, current_span.end),
                 ) {
-                    for suggestion in
-                        nested_suggestions(nuval, self.var_context.1.clone(), current_span)
+                    for suggestion in nested_suggestions(&nuval, &self.var_context.1, current_span)
                     {
                         if options.match_algorithm.matches_u8_insensitive(
                             options.case_sensitive,
@@ -140,8 +137,7 @@ impl Completer for VariableCompletion {
 
                 // If the value exists and it's of type Record
                 if let Ok(value) = var {
-                    for suggestion in
-                        nested_suggestions(value, self.var_context.1.clone(), current_span)
+                    for suggestion in nested_suggestions(&value, &self.var_context.1, current_span)
                     {
                         if options.match_algorithm.matches_u8_insensitive(
                             options.case_sensitive,
@@ -244,21 +240,21 @@ impl Completer for VariableCompletion {
 // Find recursively the values for sublevels
 // if no sublevels are set it returns the current value
 fn nested_suggestions(
-    val: Value,
-    sublevels: Vec<Vec<u8>>,
+    val: &Value,
+    sublevels: &[Vec<u8>],
     current_span: reedline::Span,
 ) -> Vec<SemanticSuggestion> {
     let mut output: Vec<SemanticSuggestion> = vec![];
-    let value = recursive_value(val, sublevels);
+    let value = recursive_value(val, sublevels).unwrap_or_else(Value::nothing);
 
     let kind = SuggestionKind::Type(value.get_type());
     match value {
         Value::Record { val, .. } => {
             // Add all the columns as completion
-            for (col, _) in val.into_iter() {
+            for col in val.columns() {
                 output.push(SemanticSuggestion {
                     suggestion: Suggestion {
-                        value: col,
+                        value: col.clone(),
                         description: None,
                         style: None,
                         extra: None,
@@ -311,56 +307,49 @@ fn nested_suggestions(
 }
 
 // Extracts the recursive value (e.g: $var.a.b.c)
-fn recursive_value(val: Value, sublevels: Vec<Vec<u8>>) -> Value {
+fn recursive_value(val: &Value, sublevels: &[Vec<u8>]) -> Result<Value, Span> {
     // Go to next sublevel
-    if let Some(next_sublevel) = sublevels.clone().into_iter().next() {
+    if let Some(next_sublevel) = sublevels.first() {
         let span = val.span();
         match val {
             Value::Record { val, .. } => {
-                for item in *val {
-                    // Check if index matches with sublevel
-                    if item.0.as_bytes().to_vec() == next_sublevel {
-                        // If matches try to fetch recursively the next
-                        return recursive_value(item.1, sublevels.into_iter().skip(1).collect());
-                    }
+                if let Some((_, value)) =
+                    val.iter().find(|(key, _)| key.as_bytes() == next_sublevel)
+                {
+                    // If matches try to fetch recursively the next
+                    recursive_value(value, &sublevels[1..])
+                } else {
+                    // Current sublevel value not found
+                    Err(span)
                 }
-
-                // Current sublevel value not found
-                return Value::nothing(span);
             }
             Value::LazyRecord { val, .. } => {
                 for col in val.column_names() {
-                    if col.as_bytes().to_vec() == next_sublevel {
-                        return recursive_value(
-                            val.get_column_value(col).unwrap_or_default(),
-                            sublevels.into_iter().skip(1).collect(),
-                        );
+                    if col.as_bytes() == *next_sublevel {
+                        let val = val.get_column_value(col).map_err(|_| span)?;
+                        return recursive_value(&val, &sublevels[1..]);
                     }
                 }
 
                 // Current sublevel value not found
-                return Value::nothing(span);
+                Err(span)
             }
             Value::List { vals, .. } => {
                 for col in get_columns(vals.as_slice()) {
-                    if col.as_bytes().to_vec() == next_sublevel {
-                        return recursive_value(
-                            Value::list(vals, span)
-                                .get_data_by_key(&col)
-                                .unwrap_or_default(),
-                            sublevels.into_iter().skip(1).collect(),
-                        );
+                    if col.as_bytes() == *next_sublevel {
+                        let val = val.get_data_by_key(&col).ok_or(span)?;
+                        return recursive_value(&val, &sublevels[1..]);
                     }
                 }
 
                 // Current sublevel value not found
-                return Value::nothing(span);
+                Err(span)
             }
-            _ => return val,
+            _ => Ok(val.clone()),
         }
+    } else {
+        Ok(val.clone())
     }
-
-    val
 }
 
 impl MatchAlgorithm {