--- a/crates/nu-cmd-extra/src/extra/filters/rotate.rs
+++ b/crates/nu-cmd-extra/src/extra/filters/rotate.rs
@@ -1,4 +1,5 @@
 use nu_engine::command_prelude::*;
+use nu_utils::Shared;
 
 #[derive(Clone)]
 pub struct Rotate;
@@ -171,7 +172,7 @@ pub fn rotate(
             let span = val.span();
             match val {
                 Value::Record { val: record, .. } => {
-                    let (cols, vals): (Vec<_>, Vec<_>) = record.into_iter().unzip();
+                    let (cols, vals): (Vec<_>, Vec<_>) = Shared::unwrap(record).into_iter().unzip();
                     old_column_names = cols;
                     new_values.extend_from_slice(&vals);
                 }