--- a/crates/nu-cmd-extra/src/extra/filters/update_cells.rs
+++ b/crates/nu-cmd-extra/src/extra/filters/update_cells.rs
@@ -1,6 +1,5 @@
 use nu_engine::{command_prelude::*, get_eval_block, EvalBlockFn};
 use nu_protocol::{ast::Block, engine::Closure, PipelineIterator};
-use nu_utils::Shared;
 use std::{collections::HashSet, sync::Arc};
 
 #[derive(Clone)]
@@ -155,8 +154,7 @@ impl Iterator for UpdateCellIterator {
                 let span = val.span();
                 match val {
                     Value::Record { val, .. } => Some(Value::record(
-                        Shared::unwrap(val)
-                            .into_iter()
+                        val.into_iter()
                             .map(|(col, val)| match &self.columns {
                                 Some(cols) if !cols.contains(&col) => (col, val),
                                 _ => (