--- a/crates/nu-command/src/filters/headers.rs
+++ b/crates/nu-command/src/filters/headers.rs
@@ -149,6 +149,7 @@ fn replace_headers(
             if let Value::Record { val: record, .. } = value {
                 Ok(Value::record(
                     record
+                        .into_owned()
                         .into_iter()
                         .filter_map(|(col, val)| {
                             old_headers