… Cargo.lock
… crates/nu-cmd-dataframe/Cargo.toml
… crates/nu-command/src/filters/drop/column.rs
… crates/nu-command/src/filters/flatten.rs
… crates/nu-command/src/filters/items.rs
… crates/nu-command/src/filters/rename.rs
… crates/nu-command/src/filters/sort.rs
… crates/nu-command/src/filters/split_by.rs
… crates/nu-command/src/filters/uniq.rs
… crates/nu-command/src/filters/values.rs
… crates/nu-command/src/formats/to/text.rs
… crates/nu-command/src/formats/to/xml.rs
… crates/nu-command/src/generators/generate.rs
… crates/nu-command/src/help/help_.rs
… crates/nu-command/src/math/utils.rs
… crates/nu-command/src/network/http/client.rs
… crates/nu-command/src/network/url/join.rs
… crates/nu-command/src/viewers/griddle.rs
… crates/nu-command/src/viewers/table.rs
… crates/nu-explore/src/nu_common/table.rs
… crates/nu-explore/src/nu_common/value.rs
… crates/nu-explore/src/pager/mod.rs
… crates/nu-plugin-test-support/src/plugin_test.rs
… crates/nu-plugin/Cargo.toml
… crates/nu-plugin/src/protocol/plugin_custom_value.rs
… crates/nu-protocol/src/config/mod.rs
… crates/nu-protocol/src/config/plugin_gc.rs
… crates/nu-protocol/src/eval_base.rs
… crates/nu-protocol/src/value/from_value.rs
… crates/nu-protocol/src/value/mod.rs