@sholderbach 2024-04-01
And for the direct comparison to #12340 

<details><summary>This rebased on main vs inserting `Record` directly into `Value`</summary>
<p>

|Name|#12340 `9057a5f`|this rebased onto `main@0cf7de5`|Relative|Note|
|-|-|-|-|-|
| load_standard_lib|12ms 250µs|12ms 320µs|0.99| |
| decoding_benchmarks|||||
|        json_decode|||||
|                (100, 5)|561µs 100ns|556µs 500ns|1.01| |
|                (10000, 15)|218ms 800µs|216ms 400µs|1.01| |
|        msgpack_decode|||||
|                (100, 5)|271µs 500ns|276µs 600ns|0.98| |
|                (10000, 15)|109ms 200µs|112ms 300µs|0.97| |
| encoding_benchmarks|||||
|        json_encode|||||
|                (100, 5)|105µs 500ns|111µs 100ns|0.95| |
|                (10000, 15)|28ms 890µs|29ms 570µs|0.98| |
|        msgpack_encode|||||
|                (100, 5)|41µs 430ns|42µs 310ns|0.98| |
|                (10000, 15)|11ms 360µs|11ms 330µs|1| |
| eval_benchmarks|||||
|        eval_default_config|2ms 895µs|2ms 537µs|1.14| |
|        eval_default_env|495µs 400ns|492µs 600ns|1.01| |
| eval_commands|||||
|        each|||||
|                1|73µs 610ns|70µs 760ns|1.04| |
|                5|217µs 600ns|174µs 500ns|1.25| |
|                10|403µs 300ns|383µs 600ns|1.05| |
|                100|3ms 598µs|3ms 211µs|1.12| |
|                1000|37ms 240µs|32ms 509µs|1.15| |
|        for_range|||||
|                1|75µs 510ns|71µs 60ns|1.06| |
|                5|202µs 800ns|212µs 600ns|0.95| |
|                10|383µs 400ns|393µs 800ns|0.97| |
|                100|3ms 638µs|3ms 581µs|1.02| |
|                1000|41ms 940µs|42ms 70µs|1| |
|        interleave|||||
|                100|1ms 229µs|1ms 286µs|0.96| |
|                1000|13ms 110µs|12ms 610µs|1.04| |
|                10000|136ms 900µs|132ms 800µs|1.03| |
|        interleave_with_ctrlc|||||
|                100|1ms 356µs|1ms 309µs|1.04| |
|                1000|13ms 100µs|13ms 90µs|1| |
|                10000|141ms|139ms 700µs|1.01| |
|        par_each_1t|||||
|                1|145µs 600ns|121µs 500ns|1.2| |
|                5|244µs 900ns|242µs 200ns|1.01| |
|                10|466µs 300ns|434µs 600ns|1.07| |
|                100|3ms 362µs|3ms 342µs|1.01| |
|                1000|31ms 680µs|33ms 30µs|0.96| |
|        par_each_2t|||||
|                1|168µs 800ns|164µs 100ns|1.03| |
|                5|246µs 700ns|243µs 600ns|1.01| |
|                10|308µs 200ns|320µs 800ns|0.96| |
|                100|1ms 900µs|1ms 905µs|1| |
|                1000|20ms 100µs|22ms 460µs|0.89| |
| parser_benchmarks|||||
|        parse_default_config_file|2ms 129µs|2ms 119µs|1| |
|        parse_default_env_file|380µs 500ns|378µs 600ns|1.01| |
| record|||||
|        create|||||
|                1|31µs 940ns|31µs 130ns|1.03| |
|                10|47µs|45µs 990ns|1.02| |
|                100|193µs 800ns|194µs 600ns|1| |
|                1000|1ms 713µs|1ms 705µs|1| |
|        flat_access|||||
|                1|17µs 980ns|18µs 430ns|0.98| |
|                10|19µs 820ns|20µs 150ns|0.98| |
|                100|30µs 820ns|24µs 840ns|1.24| |
|                1000|120µs 300ns|90µs 140ns|1.33| |
|        nest_access|||||
|                1|20µs 540ns|18µs 640ns|1.1| |
|                2|18µs 660ns|19µs 240ns|0.97| |
|                4|20µs 470ns|19µs 730ns|1.04| |
|                8|22µs 180ns|21µs 260ns|1.04| |
|                16|27µs 190ns|24µs 370ns|1.12| |
|                32|32µs 320ns|31µs 140ns|1.04| |
|                64|46µs 500ns|42µs 380ns|1.1| |
|                128|70µs 210ns|61µs 220ns|1.15| |
| table|||||
|        create|||||
|                1|35µs 490ns|37µs 460ns|0.95| |
|                10|52µs 890ns|52µs 990ns|1| |
|                100|217µs 200ns|216µs 400ns|1| |
|                1000|1ms 879µs|1ms 851µs|1.02| |
|        get|||||
|                1|26µs 970ns|29µs 130ns|0.93| |
|                10|30µs 250ns|27µs 940ns|1.08| |
|                100|59µs 340ns|42µs 990ns|1.38| |
|                1000|325µs 300ns|189µs 800ns|1.71|🟢|
|        select|||||
|                1|25µs 840ns|27µs 150ns|0.95| |
|                10|32µs 80ns|30µs 120ns|1.07| |
|                100|91µs 30ns|74µs 240ns|1.23| |
|                1000|653µs 400ns|495µs 600ns|1.32| |
||||||

</p>
</details> 

<details><summary>For reference comparing my two runs on `main@0cf7de5`</summary>
<p>

|Name|Old|New|Relative|Note|
|-|-|-|-|-|
| load_standard_lib|12ms 260µs|12ms 250µs|1| |
| decoding_benchmarks|||||
|        json_decode|||||
|                (100, 5)|557µs|559µs 700ns|1| |
|                (10000, 15)|216ms 400µs|217ms 300µs|1| |
|        msgpack_decode|||||
|                (100, 5)|273µs|277µs 900ns|0.98| |
|                (10000, 15)|111ms|112ms 200µs|0.99| |
| encoding_benchmarks|||||
|        json_encode|||||
|                (100, 5)|108µs 600ns|108µs 700ns|1| |
|                (10000, 15)|28ms 750µs|31ms 30µs|0.93| |
|        msgpack_encode|||||
|                (100, 5)|41µs 860ns|42µs 570ns|0.98| |
|                (10000, 15)|11ms 590µs|11ms 920µs|0.97| |
| eval_benchmarks|||||
|        eval_default_config|2ms 987µs|2ms 991µs|1| |
|        eval_default_env|500µs 400ns|500µs 200ns|1| |
| eval_commands|||||
|        each|||||
|                1|74µs 220ns|75µs 40ns|0.99| |
|                5|214µs 900ns|216µs|0.99| |
|                10|478µs 800ns|449µs 500ns|1.07| |
|                100|4ms 640µs|4ms 130µs|1.12| |
|                1000|54ms 260µs|53ms 230µs|1.02| |
|        for_range|||||
|                1|76µs 130ns|75µs 240ns|1.01| |
|                5|218µs|214µs 100ns|1.02| |
|                10|450µs 200ns|423µs 300ns|1.06| |
|                100|4ms 753µs|4ms 889µs|0.97| |
|                1000|56ms 450µs|57ms 700µs|0.98| |
|        interleave|||||
|                100|1ms 271µs|1ms 225µs|1.04| |
|                1000|12ms 740µs|12ms 270µs|1.04| |
|                10000|136ms 700µs|144ms 300µs|0.95| |
|        interleave_with_ctrlc|||||
|                100|1ms 338µs|1ms 327µs|1.01| |
|                1000|12ms 790µs|12ms 850µs|1| |
|                10000|139ms 200µs|147ms 300µs|0.95| |
|        par_each_1t|||||
|                1|144µs 200ns|146µs 500ns|0.98| |
|                5|252µs 700ns|258µs 899ns|0.98| |
|                10|475µs 200ns|492µs 200ns|0.97| |
|                100|3ms 985µs|3ms 483µs|1.14| |
|                1000|38ms 970µs|43ms 980µs|0.89| |
|        par_each_2t|||||
|                1|165µs 600ns|169µs|0.98| |
|                5|246µs 200ns|249µs 600ns|0.99| |
|                10|315µs 600ns|321µs 800ns|0.98| |
|                100|2ms 206µs|1ms 967µs|1.12| |
|                1000|24ms 40µs|24ms 600µs|0.98| |
| parser_benchmarks|||||
|        parse_default_config_file|2ms 167µs|2ms 16µs|1.07| |
|        parse_default_env_file|381µs 700ns|346µs|1.1| |
| record|||||
|        create|||||
|                1|32µs 80ns|30µs 210ns|1.06| |
|                10|47µs 350ns|43µs 290ns|1.09| |
|                100|193µs 900ns|176µs 200ns|1.1| |
|                1000|1ms 702µs|1ms 549µs|1.1| |
|        flat_access|||||
|                1|17µs 610ns|17µs 570ns|1| |
|                10|19µs 770ns|18µs 680ns|1.06| |
|                100|27µs 740ns|30µs 960ns|0.9| |
|                1000|119µs 200ns|119µs 300ns|1| |
|        nest_access|||||
|                1|18µs|18µs 700ns|0.96| |
|                2|18µs 890ns|18µs 410ns|1.03| |
|                4|20µs 800ns|19µs 840ns|1.05| |
|                8|24µs 180ns|22µs 810ns|1.06| |
|                16|27µs 120ns|28µs 460ns|0.95| |
|                32|36µs 190ns|36µs 280ns|1| |
|                64|49µs 210ns|49µs 200ns|1| |
|                128|76µs 130ns|75µs 640ns|1.01| |
| table|||||
|        create|||||
|                1|38µs 380ns|37µs 740ns|1.02| |
|                10|53µs 610ns|52µs 970ns|1.01| |
|                100|215µs 100ns|213µs 100ns|1.01| |
|                1000|1ms 860µs|1ms 833µs|1.01| |
|        get|||||
|                1|26µs 580ns|27µs 240ns|0.98| |
|                10|29µs 910ns|30µs 390ns|0.98| |
|                100|65µs 170ns|64µs 650ns|1.01| |
|                1000|364µs 100ns|366µs 500ns|0.99| |
|        select|||||
|                1|25µs 510ns|27µs 720ns|0.92| |
|                10|33µs 540ns|32µs 790ns|1.02| |
|                100|98µs 690ns|96µs 670ns|1.02| |
|                1000|729µs 800ns|724µs 200ns|1.01| |
||||||




</p>
</details>